add thunk support
