+++ src/App.js
@@ -2,45 +2,19 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { DebounceInput as SearchInput } from 'react-debounce-input';
 import PropTypes from 'prop-types';
-import { POSTS_URL, USERS_URL, COMMENTS_URL } from './constants';
 import PostList from './PostList';
-import { loadFromServer } from './api';
+import { loadPostsFromServer } from './store/posts';
 import './App.css';
 import { getSearchQuery, getPosts, getIsLoading } from './store';
-import { setPosts } from './store/posts';
 import { setSearchQuery } from './store/searchQuery';
-import { setIsLoading } from './store/isLoading';
 
 const App = (
-  { posts,
-    isLoading,
+  { posts, isLoading, searchQuery,
     // eslint-disable-next-line no-shadow
-    searchQuery, setPosts, setSearchQuery, setIsLoading }
+    setSearchQuery, loadPostsFromServer }
 ) => {
-  const normalizePosts = (postsList, usersList, commentsList) => postsList
-    .map(post => ({
-      ...post,
-      comments: commentsList.filter(comment => comment.postId === post.id),
-      user: usersList.find(user => user.id === post.userId),
-    }));
-
   const loadPosts = async() => {
-    setIsLoading(true);
-
-    const [postsList, usersList, commentsList] = await Promise.all([
-      loadFromServer(POSTS_URL),
-      loadFromServer(USERS_URL),
-      loadFromServer(COMMENTS_URL),
-    ]);
-
-    const postsData = normalizePosts(
-      postsList,
-      usersList,
-      commentsList,
-    );
-
-    setPosts(postsData);
-    setIsLoading(false);
+    await loadPostsFromServer();
   };
 
   const filterPosts = (e) => {
@@ -89,14 +63,13 @@ const mapStateToProps = state => ({
 });
 
 export default connect(mapStateToProps, {
-  setPosts, setSearchQuery, setIsLoading,
+  setSearchQuery, loadPostsFromServer,
 })(App);
 
 App.propTypes = {
   posts: PropTypes.arrayOf(PropTypes.object).isRequired,
   isLoading: PropTypes.bool.isRequired,
   searchQuery: PropTypes.string.isRequired,
-  setPosts: PropTypes.func.isRequired,
   setSearchQuery: PropTypes.func.isRequired,
-  setIsLoading: PropTypes.func.isRequired,
+  loadPostsFromServer: PropTypes.func.isRequired,
 };
