+++ src/store/posts.js
@@ -1,3 +1,7 @@
+import { loadFromServer } from '../api';
+import { setIsLoading } from './isLoading';
+import { COMMENTS_URL, POSTS_URL, USERS_URL } from '../constants';
+
 const SET_POSTS = 'setPosts';
 const DELETE_POST = 'deletePost';
 const DELETE_COMMENT = 'deleteComment';
@@ -17,6 +21,32 @@ export const deleteComment = (postId, commentId) => ({
   postId,
   commentId,
 });
+
+export const loadPostsFromServer = () => async(dispatch) => {
+  const normalizePosts = (posts, users, comments) => posts
+    .map(post => ({
+      ...post,
+      comments: comments.filter(comment => comment.postId === post.id),
+      user: users.find(user => user.id === post.userId),
+    }));
+
+  dispatch(setIsLoading(true));
+
+  const [postsList, usersList, commentsList] = await Promise.all([
+    loadFromServer(POSTS_URL),
+    loadFromServer(USERS_URL),
+    loadFromServer(COMMENTS_URL),
+  ]);
+
+  const postsData = normalizePosts(
+    postsList,
+    usersList,
+    commentsList,
+  );
+
+  dispatch(setPosts(postsData));
+  dispatch(setIsLoading(false));
+};
 
 const postsReducer = (posts = [], action) => {
   switch (action.type) {
